Moved accordion script into component folder
