+++ assets/js/components.js
@@ -9,3 +9,4 @@ import '../../components/02-form-elements/textarea/character-limit.js';
 import '../../components/definitions/definitions.js';
 import '../../components/02-form-elements/input/input--password';
 import '../../components/timeout/timeout.js';
+import '../../components/accordion/accordion.js';
